pyproject.toml
[tool.poetry]
name = "django-test-task"
version = "0.1.0"
description = ""
authors = ["Hryhorenko Vitalii <hryhorenko.vitalii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
django-filter = "^22.1"
django-celery-results = "^2.4.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
